Split association classes into separate files
