+++ lib/rom/sql/association/one_to_many.rb
@@ -1,0 +1,8 @@
+module ROM
+  module SQL
+    class Association
+      class OneToMany < OneToOne
+      end
+    end
+  end
+end
